requirements.txt
openai>=0.27.0
python-dotenv>=0.15.0
pyperclip>=1.8.2